+++ sync-for-payables/RELEASES.md
@@ -79,3 +79,13 @@ Based on:
 - [typescript v1.4.0] sync-for-payables
 ### Releases
 - [NPM v1.4.0] https://www.npmjs.com/package/@codat/sync-for-payables/v/1.4.0 - sync-for-payables
+
+## 2023-10-02 00:04:58
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Sync-Payables.yaml
+- Speakeasy CLI 1.92.3 (2.143.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v1.5.0] sync-for-payables
+### Releases
+- [NPM v1.5.0] https://www.npmjs.com/package/@codat/sync-for-payables/v/1.5.0 - sync-for-payables